==> scripts/old_scripts/02-data-explore.Rmd <==
---
title: "Anthropause initial sound pressure and weather data exploration"
author: "Philina English"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
if (!require(tidyverse)) install.packages("tidyverse"); library(tidyverse)
if (!require(lubridate)) install.packages("lubridate"); library(lubridate)
if (!require(EnvStats)) install.packages("EnvStats") 
if (!require(devtools)) install.packages("devtools")
if (!require(gfplot)) devtools::install_github("pbs-assess/gfplot")
if (!require(nlme)) install.packages("nlme");library(nlme)
if (!require(mgcv)) install.packages("mgcv")
if (!require(gamm4)) install.packages("gamm4")
if (!require(effects)) install.packages("effects")

theme_set(gfplot:::theme_pbs())

# all sound pressure data pooled by hr
# splhr <- readRDS("wdata/splbyhour.rds") 

# sound pressure and weather data 
daily <- readRDS(here::here("wdata/trimmed_daily_weather.rds"))
hourly <- readRDS(here::here("wdata/trimmed_hourly_weather.rds"))

hourly$hrseq <- ((as.numeric(hourly$datehr) - min(as.numeric(hourly$datehr))+ 3600)/3600)  
hourly$hrseq
hourly$hrseq.f <- as.factor((as.numeric(hourly$datehr) - min(as.numeric(hourly$datehr))+ 3600)/3600)  
hourly$hr.c <- (hourly$hr)-12
hourly$spl50

hourly2 <- hourly %>% select(-wdir) %>% drop_na()
hourly2$year.f <- as.factor(hourly2$year)

hourly2$rca <-as.factor(hourly2$rca) 
hourly2$year <-as.factor(hourly2$year) 
hourly2$position <-as.factor(hourly2$position) 
hourly2$phase <- hourly2$period
hourly2$phase[hourly2$phase == "early"] <- "I"
hourly2$phase[hourly2$phase == "pre-ferry"] <- "II-pre-ferry"
hourly2$phase[hourly2$phase == "with-ferry"] <- "II-with-ferry"

hourly2 <- hourly2 %>% arrange(datehr) %>%
  group_by(rca, year, position) %>%
  mutate(
    day.drop = (as.numeric(date) - min(as.numeric(date)) + 3600*24)/3600*24,
    hr.drop = (as.numeric(datehr) - min(as.numeric(datehr)) + 3600)/3600
    ) %>% 
  ungroup() %>%
  group_by(year, phase) %>% 
  mutate(
    hr.period = (as.numeric(datehr) - min(as.numeric(datehr)) + 3600)/3600 
    ) %>%
  ungroup() %>%
  mutate(
    period = as.factor(paste(phase, year)),
    phase = as.factor(phase)
    )
  
# glimpse(hourly2)
```

```{r eval = FALSE, include = FALSE}
daily %>% 
  ggplot(aes(y = wspeed, x = as.factor(year), color = as.factor(year))) +
  geom_violin() +
  geom_point(alpha =0.2) +
  facet_grid(rca~period, scales = "free")

hourly %>% 
  ggplot(aes(y = wspeed, x = as.factor(year), color = as.factor(year))) +
  geom_violin() +
  geom_jitter(alpha =0.2) +
  facet_grid(rca~period, scales = "free")
```

```{r eval = FALSE, include = FALSE}
# daily %>% glimpse()
hourly %>% glimpse()
```

```{r}
hist(hourly$spl50, breaks = 20)
```

```{r}
hourly %>% 
  ggplot(aes(y = spl50, x = wspeed, color = as.factor(year), fill = as.factor(year))) +
  # geom_smooth(method = "lm") +
  geom_smooth(method= "lm", formula = y~x + I(x^2)) +
  geom_point(alpha = 0.2, size = 2) +
  facet_wrap(~rca) +
  xlab ("Wind speed") +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  ylab("Median hourly sound pressure") +
  ggtitle("Soundscape ('in' is at 12m depth; 'out' is only 10m and closer to shore)")

```


## Check for effect of wind speed at quietest time of night
Noisier at RCA out... maybe because closer to shore?

```{r}
# wind speed at night 
hourly %>% filter(
  hr < 4 & hr > 0
) %>%
  ggplot(aes(y = spl50, x = wspeed, color = as.factor(year), fill = as.factor(year))) +
  # geom_smooth(method = "lm") +
  geom_smooth(method= "lm", formula = y~x + I(x^2)) +
  geom_point(alpha = 0.2, size = 2) +
  facet_wrap(~rca) +
  xlab ("Wind speed") +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  ylab("Median hourly sound pressure") +
  ggtitle("Soundscape in wee hours of night")

```

### Wind speed under 20 km/hr seems acceptable 
```{r}
# wind speed under 20 seems acceptable 
hourly %>% filter( hr < 4 & hr > 0 &
  wspeed < 20 # ~ 10 knots
  ) %>%
  ggplot(aes(y = spl50, x = wspeed
    , color = rca, fill = rca
    #, color = as.factor(year), fill = as.factor(year)
    )) +
  geom_smooth(method = "lm") +
  # geom_smooth(method= "lm", formula = y~x + I(x^2)) + # doesn't work for under 20 km
  geom_point(alpha = 0.2, size = 2) +
  # facet_grid(rca~period, scales = "free")+
  xlab ("Wind speed") +
  scale_fill_manual(name= "RCA",values = c("#00AFBB", "#E7B800" ))+ #"#E7B800",
  scale_color_manual(name= "RCA",values = c("#00AFBB", "#E7B800"))+
  ylab("Median hourly sound pressure") +
  ggtitle("Soundscape in wee hours at night when wind speed <20 km/hr ")
```

```{r}
hourly %>% filter(
  (hr < 4 | hr > 10), # at night only
  wspeed < 20 # ~ 5 knots
  ) %>%
  ggplot(aes(y = spl50, x = wspeed, color = as.factor(year), fill = as.factor(year))) +
  geom_smooth(method = "lm") +
  geom_point(alpha = 0.2, size = 2) +
  facet_grid(rca~period, scales = "free")+
  xlab ("Wind speed") +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  ylab("Median hourly sound pressure") +
  ggtitle("Soundscape all night when wind speed <20 km/hr")
```

## What time of day is noisiest?
```{r }
hourly %>% filter(
  wspeed < 20 # ~ 10 knots
) %>%
  ggplot(aes(y = spl50, x = hr, color = as.factor(year), fill = as.factor(year))) +
  geom_smooth(method= "lm", formula = y~x + I(x^2)) +
  geom_point(alpha = 0.2, size = 2) +
  facet_grid(rca~period, scales = "free") +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  xlab ("Time of day") +
  ylab("Median hourly sound pressure") +
  ggtitle("Changes through the day when wind speed <20 km/hr")
```

## Boxplots
```{r}
# boxplot
hourly %>% filter(
  # hr >= 9 & hr <= 17, # between 9-5; busiest
  # hr < 4 & hr > 2, # between 2-4 am; quietest 
  wspeed < 20
) %>%
  ggplot(aes(y = spl50, x = as.factor(period), color = as.factor(year))) +
  # geom_violin() +
  geom_boxplot() +
  EnvStats::stat_n_text(y.pos = 32, alpha = 0.5) + 
  # geom_jitter(alpha=0.2) +
  facet_wrap(~rca, nrow = 2) +
  # facet_wrap(~grp, nrow = 2) +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  xlab("Time period") +
  ylab("Median hourly sound pressure") +
  ggtitle("Soundscape differences with COVID when wind < 20 km/hr")
```

```{r eval= F}
# boxplot
hourly %>% filter(
  # hr >= 9 & hr <= 17, # between 9-5; busiest
  hr < 4 & hr > 0, # between 2-4 am; quietest
  wspeed < 10
) %>%
  ggplot(aes(y = spl50, x = as.factor(period), color = as.factor(year), fill = as.factor(year))) +
  # # geom_violin() +
  #  geom_violin(aes(y = spl95, x = as.factor(period), colour = as.factor(year)), inherit.aes = F) +
  #  geom_violin(aes(y = spl05, x = as.factor(period), colour = as.factor(year)), inherit.aes = F) +
   geom_boxplot(alpha = 0.5) +
  EnvStats::stat_n_text(y.pos = 28, alpha = 0.5) + 
  # geom_jitter(alpha=0.2) +
  facet_wrap(~rca, nrow = 2) +
  # facet_wrap(~grp, nrow = 2) +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  # scale_fill_discrete(name= "Year") +
  # scale_colour_discrete(name= "Year") + 
  xlab("Time period") +
  ylab("Median hourly sound pressure") +
  ggtitle("Soundscape nighttime differences with COVID when wind < 10 km/hr")
```

## Busiest time of day
```{r}
# boxplot
hourly %>% filter(
  hr >= 9 & hr <= 17, # between 9-5; busiest
  # hr < 4 & hr > 2, # between 2-4 am; quietest 
  wspeed < 20
) %>%
  ggplot(aes(y = spl50, x = as.factor(period), color = as.factor(year))) +
  # geom_violin() +
  geom_boxplot( outlier.alpha = 0) +
  EnvStats::stat_n_text(y.pos = 28, alpha = 0.5) + 
  geom_point(position = position_jitterdodge(), alpha=0.2) +
  facet_wrap(~rca, nrow = 2) +
  # facet_wrap(~grp, nrow = 2) +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  xlab("Time period") +
  ylab("Median hourly sound pressure") +
  ggtitle("Soundscape daytime differences with COVID when wind < 20 km/hr")
```

## Explore the effect of wind direction at both sites
All directions also noisier at RCA out
But hopefully we can ignore direction if calm 

```{r include = F, eval=F}
hourly %>% filter(
  wspeed > 20 # ~ 5 knots
) %>% 
  # mutate(wdirN = wdir-27, wdirN = if_else(wdirN == - 18, 0, wdirN)) %>%
  ggplot(aes(y = spl50, x = wdir)) +
  geom_smooth(method= "lm", formula = y~x + I(x^2)) +
  geom_point(alpha = 0.2, size = 2) +
  facet_grid(rca~period, scales = "free") +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  # xlab ("Time of day") +
  ylab("Median hourly sound pressure") +
  ggtitle("Changes with wind direction")
```

### check wind directions at night and early morning
```{r}
# N or NW winds
hourly %>% filter(
  #wspeed < 10, 
  hr < 7,
  wdir == 1 |  wdir >= 32) %>% # effects both equally
  ggplot(aes(y = spl50, x = wspeed, color = as.factor(year), group = grp)) +
  geom_smooth(method = "lm") +
  geom_point(alpha = 0.2, size = 2) +
  facet_grid(~rca, scales = "free") +  
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  ggtitle("N or NW winds")

# NE or E winds are rare
hourly %>% filter(  
  hr < 7,
  wdir > 1 & wdir < 13) %>%
  ggplot(aes(y = spl50, x = wspeed, color = as.factor(year), group = grp)) +
  geom_smooth(method = "lm") +
  geom_point(alpha = 0.2, size = 2) +
  facet_grid(~rca, scales = "free") +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  ggtitle("NE or E winds are rare")

# SE winds
hourly %>% filter(#wspeed < 10, 
  hr < 7,
  wdir >= 13 & wdir < 18) %>% # appears to effect RCA in most
  ggplot(aes(y = spl50, x = wspeed, color = as.factor(year), group = grp)) +
  geom_smooth(method = "lm") +
  geom_point(alpha = 0.2, size = 2) +
  facet_grid(~rca, scales = "free") +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  ggtitle("SE winds")

# SW OR W winds
hourly %>% filter(
  #wspeed < 10, 
  hr < 7,
  wdir >= 18 & wdir < 32) %>% # appears to effect RCA in most
  ggplot(aes(y = spl50, x = wspeed, color = as.factor(year), group = grp)) +
  geom_smooth(method = "lm") +
  geom_point(alpha = 0.2, size = 2) +
  facet_grid(~rca, scales = "free") +
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  ggtitle("SW OR W winds")
```


```{r include=F, eval=F}
# ANOVA
m <- aov(spl50 ~ 
    position + # unique intercepts for each location 
    period * as.factor(year) +# time variables interacting?
    period * rca # site difference might vary seasonally?
  , data = filter(hourly, wspeed < 20))
summary(m)
```

```{r include=F, eval=F}
m <- aov(spl50 ~ 
    period * as.factor(year) + # year and postition are confounded
    period * rca # site difference might vary seasonally?
  , data = filter(hourly, wspeed < 20))
summary(m)
TukeyHSD(m)
```


```{r include=F, eval=F}
# ANCOVA
# add daily curve 
m <- lm(spl50 ~ poly(hr, 2) + period * as.factor(year) + period * rca, 
  data = filter(hourly, wspeed < 20))
summary(m)
```

```{r include=F, eval=F}
# but the shape of daily curve does interact with year and period
m <- lm(spl50 ~ poly(hr, 2) * period * as.factor(year) + period * rca, 
  data = filter(hourly, wspeed < 20))
summary(m)
```

```{r include=F, eval=F}
# so extract noisiest period only 
m <- lm(spl50 ~ period * as.factor(year) + period * rca
  , data = filter(hourly, wspeed < 10 & hr >= 9 & hr <= 17))
summary(m)
```

Time continuous plot:
In full data set there is temporal autocorrelation that we would need to control for
```{r}
hourly %>% 
ggplot(aes(y=spl50, x=mdh, color=as.factor(year), group=grp))+
  geom_point(alpha = 0.5, size = 0.5)+
  geom_line(alpha = 0.5)+
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  # coord_cartesian(ylim = c(90, 120))+
  facet_grid(rca~period, scales="free") +
  xlab("Time")
```

Still quite a bit in the daytime low wind version 
```{r}
hourly %>% filter( 
  wspeed < 20 #& hr >= 9 & hr <= 17
  ) %>%
ggplot(aes(y=spl50, x=mdh, color=as.factor(year), group=grp))+
  geom_point(alpha = 0.8)+
  geom_line(alpha = 0.2)+
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  # coord_cartesian(ylim = c(90, 120))+
  facet_grid(rca~period, scales="free") +
  xlab("Time")
```

```{r}
hourly %>% filter( 
  # wspeed < 20 #& hr >= 9 & hr <= 17
  ) %>%
ggplot(aes(y=wspeed, x=mdh, color=as.factor(year), group=grp))+
  geom_point(alpha = 0.8)+
  geom_line(alpha = 0.2)+
  scale_fill_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+ #"#E7B800",
  scale_color_manual(name= "Year",values = c("#FC4E07", "#00AFBB"))+
  # coord_cartesian(ylim = c(90, 120))+
  facet_grid(rca~period, scales="free") +
  xlab("Time")
```


# TRY WITH TEMPORAL AR1 using glmmTMB

### first for winds less than 20 km per hour
```{r}
hourly2$site <- paste0("RCA", hourly2$rca)
hourly2$deployment <-  hourly2$position
hourly2$time <- hourly2$hrseq.f
hourly2$timeofday <- hourly2$hr.c
hourly2$wind_speed <- hourly2$wspeed

hourlySPLdata <- hourly2

mAR120 <- glmmTMB::glmmTMB(spl50 ~ phase * year + poly(timeofday, 2) + site + 
    (1|deployment) + 
    # us(time-1|position), #crashed computer !
    ar1(time-1|deployment),
  data = filter(hourlySPLdata, wind_speed < 20), family=gaussian)

summary(mAR120)
```
```{r}
mAR120$frame$resid <- residuals(mAR120, type = "pearson")
mAR120$frame$est <- predict(mAR120)

ggplot(mAR120$frame, aes(est, resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(phase~year)

ggplot(mAR120$frame, aes(`poly(timeofday, 2)`[,1], resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(site~year, scales = "free")

acf(mAR120$frame$resid, resType = "pearson")

plot(effects::Effect(c("timeofday", "site"), mAR120, residuals=TRUE), partial.residual=list(lty="dashed", col= "black"))
plot(effects::Effect(c("timeofday", "year"), mAR120, residuals=TRUE), partial.residual=list(lty="dashed", col= "black"))
plot(effects::Effect(c("timeofday", "phase"), mAR120, residuals=TRUE), partial.residual=list(lty="dashed", col= "black"))
```

```{r}

din19 <- filter(mAR120$frame, site == "RCAin" & year == 2019 & phase == "I") %>% arrange(time) 
dout19 <- filter(mAR120$frame,  site == "RCAout" & year == 2019 & phase == "I") %>% arrange(time) 


din <- filter(mAR120$frame,  site == "RCAin" & year == 2020 & phase == "I") %>% arrange(time) 
dout <- filter(mAR120$frame,  site == "RCAout" & year == 2020 & phase == "I") %>% arrange(time) 

acf(din19$resid, lag = 20)
acf(dout19$resid, lag = 20)

acf(din$resid, lag = 20)
acf(dout$resid, lag = 20)

mAR120$frame$resid <- residuals(mAR120, type = "pearson")
d <- filter(mAR120$frame) %>% arrange(time) 
acf(d$resid, lag = 40)
```

### include all data so as not to violate AR1 assumptions

```{r}
hourly2$site <- paste0("RCA", hourly2$rca)
hourly2$deployment <-  hourly2$position
hourly2$time <- hourly2$hrseq.f
hourly2$timeofday <- hourly2$hr.c
hourly2$log_wind_speed <- log(hourly2$wspeed+1)
hourly2$wind_speed <- (hourly2$wspeed)

hourlySPLdata <- hourly2
hourlySPLdata$log_spl50 <- log(hourlySPLdata$spl50+1)
hourlySPLdata$spl50_raw <- hourlySPLdata$spl50 + 87.7654

mAR1 <- glmmTMB::glmmTMB(spl50_raw ~ year * phase + 
    # poly(timeofday, 2) + 
    # poly(timeofday, 2) * site + 
    poly(wind_speed, 2) * site +
    # log_wind_speed * site +
    (1|deployment) + 
    ar1(time-1|deployment),
  # REML = T,
  data = hourlySPLdata, 
  family=gaussian)
  # family=gaussian(link = "log"))

summary(mAR1)
car::Anova(mAR1, 
  type = "III"
  # component = "zi"
  )

print(mAR1)  
```



```{r}
# min + est for 2019
est2019 <- 87.7654 + 14.9240
# [1] 102.6894

# min + est for 2019 + change to 2020
est2020 <- 87.7654 + 14.9240 -8.0439
# [1] 94.6455
# % change in dB
((est2020-87.7654)/(est2019-87.7654))-1


# conversion to pressure is 10^(SPL/10)
(10^(est2020/10))/(10^(est2019/10))
# [1] 0.1569
(10^(est2020/10))/(10^(est2019/10))-1
# [1] 0.8431

10^(est2020/10)
10^(est2019/10)

# simpler version
(10^((14.9240-8.0439)/10))/(10^(14.9240/10))-1

10^((14.9240-8.0439)/10)
10^(14.9240/10)

hist(hourlySPLdata$spl50)
hist(hourlySPLdata$log_spl50)
# try making response in log space but it doesn't change much
87.7654 + exp(2.67806)
87.7654 + exp(2.67806-0.60768)

# conversion to pressure is 10^(SPL/10)
# (10^(95.69324/10))/(10^(102.3222/10))
1-(10^(95.69324/10))/(10^(102.3222/10))
```

These behaviour changes have resulted in a reduction in difference between mean hourly sound pressure levels and the minimum level ever recorded at either of these locations from close to 14 dB to less than 7 dB during phase I of lockdown. This effect is much smaller in phase II, even before the ferry resumes, likely do to good weather and a resumption, or even an increase over 2019, in recreational activity.

```{r}
mAR1$frame$resid <- residuals(mAR1, type = "pearson")
mAR1$frame$est <- predict(mAR1)

ggplot(mAR1$frame, aes(est, resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(phase~year)

ggplot(mAR1$frame, aes(`poly(timeofday, 2)`[,1], resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(~deployment, scales = "free")

ggplot(mAR1$frame, aes(`poly(wind_speed, 2)`[,1], resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(~deployment, scales = "free")

ggplot(mAR1$frame, aes(`poly(wind_speed, 2)`[,2], resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(~deployment, scales = "free")

# acf(mAR1$frame, resType = "pearson")

effects::Effect(c("site", "year"), mAR1, residuals=TRUE)

plot(effects::Effect(c("year", "site"), mAR1, residuals=TRUE), partial.residual=list(lty="dashed", col= "black"))
plot(effects::Effect(c("timeofday", "year"), mAR1, residuals=TRUE), partial.residual=list(lty="dashed", col= "black"))
plot(effects::Effect(c("wind_speed", "phase"), mAR1, residuals=TRUE), partial.residual=list(lty="dashed", col= "black"))
# plot(effects::Effect(c("log_wind_speed", "site"), mAR1, residuals=TRUE), partial.residual=list(lty="dashed", col= "black"))

# plot(effects::Effect(c("wind_speed", "deployment"), mAR1, residuals=TRUE), partial.residual=list(lty="dashed", col= "black"))
plot(effects::Effect(c("wind_speed", "site"), mAR1, residuals=TRUE), partial.residual=list(lty="dashed", col= "black"))
```

```{r}
din19 <- filter(mAR1$frame, site == "RCAin" & year == 2019 & phase == "I") %>% arrange(time) 
dout19 <- filter(mAR1$frame,  site == "RCAout" & year == 2019 & phase == "I") %>% arrange(time) 

din <- filter(mAR1$frame,  site == "RCAin" & year == 2020 & phase == "I") %>% arrange(time) 
dout <- filter(mAR1$frame,  site == "RCAout" & year == 2020 & phase == "I") %>% arrange(time) 

acf(din19$resid, lag = 20)
acf(dout19$resid, lag = 20)

acf(din$resid, lag = 20)

dout <- filter(mAR1$frame,  site == "RCAout" & year == 2020 & phase == "I") %>% arrange(time) 
acf(dout$resid, lag = 20)

mAR1$frame$resid <- residuals(mAR1, type = "pearson")
d <- filter(mAR1$frame) %>% arrange(time) 
acf(d$resid, lag = 40)
```


# TRY NLME approach
### uses data variable, but not sure how
### adding different lags not changing anything

```{r }
mnlme <- lme(spl50 ~ phase * year + poly(wind_speed,2)*rca, # + #poly(hr.c, 2)*rca,
    random = list(~1|deployment), 
             data = hourlySPLdata)
summary(mnlme)
ACF(mnlme,
    form = ~ hrseq.f|position)  

ACF(mnlme,
    form = ~ datehr|position)
```


```{r eval = F}
mGLS1 <- lme(spl50 ~ phase * year + 
    poly(timeofday, 2) + poly(wind_speed,2)
  , random = list(~1|deployment), 
  correlation = corAR1(form = ~ datehr|deployment,  value = 0.49), 
  # correlation = corARMA(form = ~ datehr|deployment, p=1, q = 1),
  # method="REML",
  data = hourlySPLdata
  )

summary(mGLS)
summary(mGLS1)

mGLS$data$resid <- residuals(mGLS, type = "normalized")
mGLS$data$est <- predict(mGLS)

saveRDS(mGLS, file = "models/spl50-gls-no-rca.rds")
```

AIC 27417.22

```{r eval = F}
mGLSrca <- lme(spl50 ~ phase * year + 
    # poly(timeofday, 2) + 
    poly(wind_speed,2)* rca
    # poly(hr.c, 2) * rca + poly(wspeed, 2) * rca
  , random = list(~1|deployment), 
  correlation = corAR1(form = ~ datehr|deployment
    ,  value = 0.72 # value estimated with glmmTMB
    # ,  value = 0.52 # .49 if timeofday, both from ACF function
    ),
  # correlation = corARMA(form = ~ datehr|deployment, p=3, q = 1),
  method="ML",
  # method="REML",
  data = hourlySPLdata
  )

summary(mGLSrca)

mGLSrca$data$resid <- residuals(mGLSrca, type = "normalized")
mGLSrca$data$est <- predict(mGLSrca)

saveRDS(mGLSrca, file = "models/spl50-gls-w-rca2.rds") 
```

```{r}
mGLSrca <- readRDS("models/spl50-gls-w-rca2.rds")
summary(mGLSrca)
```

AIC 
only wind interacts with rca: 27384.65	
both interact with rca: 27360.28

```{r}
coef(summary(mGLSrca))
```

all data and both wind and hr interact with rca
```{r}
mGLSrca$data$resid <- residuals(mGLSrca, type = "normalized")
ggplot(filter(mGLSrca$data), aes(est, resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  # facet_grid(rca~year.f, scales = "free")
  facet_grid(year.f~rca)
```


# Models without winds > 20 km 
better with no effect of wind at these levels

```{r}
mnlme2 <- lme(spl50 ~ phase * year + 
    poly(hr.c, 2)*rca,
    random = list(~1|position), 
             data = filter(hourly2, wspeed < 20
               ))
summary(mnlme2)
ACF(mnlme2,
    form = ~ datehr|position)
```


```{r eval = F}
mGLS20 <- nlme::lme(spl50 ~ phase * year + #wspeed + 
  poly(hr.c, 2) + rca,
  # poly(hr.c, 2) * rca,
  random = list(~1|position), 
  # correlation = corAR1(form = ~ datehr|position
  # ,  value = 0.38
  # ),
   # correlation = corARMA(value = c(0.38, 0.24),
   #  form = ~ datehr|position, p=2, q = 0),
  correlation = corARMA(value = c(0.38, 0.24, 0.16, 0.3),
    form = ~ datehr|position, p=3, q = 1),
  # method="ML",
  method="REML",
  data = filter(hourly2, wspeed < 20)
  )


mGLS20$data$resid <- resid(mGLS20, type = "normalized")
mGLS20$data$est <- predict(mGLS20)

# saveRDS(mGLS20, file = "models/spl50-gls-20km-no-wind-p-2.rds")
saveRDS(mGLS20, file = "models/spl50-gls-20km-no-wind-p3-q1.rds")
# saveRDS(mGLS20, file = "models/spl50-gls-20km-no-wind.rds")
# saveRDS(mGLS20, file = "models/spl50-gls-20km-linear.rds")
```

```{r}
# mGLS20 <- readRDS("models/spl50-gls-20km-no-wind-arma.rds")
# mGLS20 <- readRDS("models/spl50-gls-20km-linear.rds")

 mGLS20 <- readRDS("models/spl50-gls-20km-no-wind.rds")
 # mGLS20 <- readRDS("models/spl50-gls-20km-no-wind-p3-q1.rds")
summary(mGLS20)
```
AIC
just main effect of rca and arma: 14960.25	
with wind: 14954.63	
no wind : 14946.54
no wind arma : 14948.54	

check RE
no wind , p =3 : 14962.25
no wind , p =2 : 14960.25
no wind , p =1 : 14958.25	

### try also removing RCA but seems better to keep it
```{r eval = F}
mGLS20b <- lme(spl50 ~ period * year.f + 
     # wspeed * rca + 
     poly(hr.c, 2) #* rca
  , 
     random = list(~1|position), 
  correlation = corAR1(form = ~ datehr|position
    # ,  value = 0.43
    ), 
  # method="REML",
  data = filter(hourly2, wspeed < 20)
  )
mGLS20b$data$resid <- residuals(mGLS20b)
mGLS20b$data$est <- predict(mGLS20b)

saveRDS(mGLS20b, file = "models/spl50-gls-20km-wout-rca.rds")
```

```{r eval = F}
mGLS20b <- readRDS("models/spl50-gls-20km-wout-rca.rds")
summary(mGLS20b)
```
AIC
time quad universal: 14965.05	
vs interact with RCA: 14946.54

```{r}
round(coef(summary(mGLS20)), digits = 4)
```

wind less than 20 km and RCA * hr
```{r}
plot(effects::Effect(c("hr.c", "rca"), mGLS20, residuals=TRUE), partial.residual=list(lty="dashed"))
plot(effects::Effect(c("hr.c", "year"), mGLS20, residuals=TRUE), partial.residual=list(lty="dashed"))
plot(effects::Effect(c("hr.c", "phase"), mGLS20, residuals=TRUE), partial.residual=list(lty="dashed"))
 
mGLS20$data$resid <- resid(mGLS20, type = "normalized")
mGLS20$data$resid2 <- resid(mGLS20)
ggplot(mGLS20$data, aes(est, resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(rca~year, scales = "free")
ggplot(mGLS20$data, aes(est, resid2)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(rca~year, scales = "free")
  

ggplot(mGLS20$data, aes(wspeed, resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(rca~year, scales = "free")
ggplot(mGLS20$data, aes(hr.c, resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(rca~year, scales = "free")

plot(mGLS20, resid(., type = "response") ~ fitted(.) | year, abline = 0)
plot(mGLS20, resid(., type = "normalized") ~ fitted(.) | year, abline = 0)
plot(mGLS20, resid(., type = "normalized") ~ fitted(.) | rca, abline = 0)
plot(mGLS20, resid(., type = "normalized") ~ fitted(.) | wspeed, abline = 0)

ggplot(mGLS20$data, aes(mdh, resid)) + 
  geom_point(alpha = 0.2)  +
  # geom_line() + 
  facet_grid(rca~year.f, scales = "free")

mGLS20$data %>% filter (month == 4) %>%
ggplot(., aes(datehr, resid)) + 
  geom_point(alpha = 0.2)  +
  geom_line() +
  facet_grid(rca~year.f, scales = "free")
```

```{r eval=F}
din19 <- filter(mGLS20$data, rca == "in" & year == 2019 & phase == "I") %>% arrange(datehr) 
dout19 <- filter(mGLS20$data, rca == "out" & year == 2019 & phase == "I") %>% arrange(datehr) 


din <- filter(mGLS20$data, rca == "in" & year == 2020 & phase == "I") %>% arrange(datehr) 
dout <- filter(mGLS20$data, rca == "out" & year == 2020 & phase == "I") %>% arrange(datehr) 

acf(din19$resid, lag = 20)
acf(dout19$resid, lag = 20)

acf(din$resid, lag = 20)
acf(dout$resid, lag = 20)


d <- filter(mGLS20$data) %>% arrange(rca, datehr) 

acf(d$resid, lag = 40)
mGLS20$data$resid2 <- residuals(mGLS20, type = "normalized")

plot(ACF(mGLS20, resType = "normalized", group = "year"))
```



without RCA
```{r eval = F}
mGLS20b$data$resid <- residuals(mGLS20b, type = "normalized")
ggplot(filter(mGLS20b$data), aes(est, resid)) + geom_point(alpha = 0.2)  + geom_smooth() + 
  facet_grid(year.f~rca, scales = "free")
  # facet_grid(period~year.f)
```




```{r eval = F}
hourly2$time <- as.numeric(hourly2$datehr)
min(hourly2$time)
hourly2$time <- hourly2$time - min(hourly2$time)

gam1 <- mgcv::gam(spl50 ~ 
    # s(h2, bs = "cr", k =24) +
    s(h2, bs = "ps", k = 24) +
    # s(time)+ 
    rca + as.factor(year) * period,  
    data = hourly)

summary(gam1)

plot(gam1, shade = TRUE)
mgcv::gam.check(gam1)
# mgcv::vis.gam(gam1)
```



```{r eval = F}
gam2 <- gamm4::gamm4(
# gam2 <-mgcv::gam(
  spl50 ~
    # s(hr.drop, bs="ts", by = position) +
    s(hr.drop, bs = "fs", by = position) + # fs gives intercept
    # s(hr.c, bs = "cc", k = 6, by = rca) +
    s(wspeed, bs = "fs", by = rca) +
    year * phase,  
  random = ~ (1|position) + (1|rca),
  data = hourly2)

# saveRDS(gam2, file = "Gamm-hr-day-position.rds")
saveRDS(gam2, file = "Gamm-hr-position-hr-rca.rds")

summary(gam2$gam)
summary(gam2$mer)

# summary(gam2)
```

```{r eval = F}
plot(gam2$gam, shade = TRUE, ylim = c(-6,8)) 

plot(gam2$mer)

hourly2$gam_resids <- gam2$gam$residuals
hourly2$gam_resids2 <- residuals(gam2$mer, type = "pearson")
 
mgcv::gam.check(gam2$gam)

d <- filter(hourly2, position == "drop4") %>% arrange(position, datehr) 
acf(d$gam_resids, lag = 30)
acf(d$gam_resids2, lag = 30)
```


wind less than 20 km and RCA * hr
```{r eval = F}
hourly2$gamresid <- gam2$gam$residuals
hourly2$gamest <-gam2$gam$fitted.values

ggplot(hourly2, aes(gamest, gamresid)) + geom_point(alpha = 0.2)  + 
  geom_smooth() + 
  facet_grid(year.f~rca, scales = "free")
```


```{r eval = F}
ggplot(hourly2, aes(mdh, gamresid)) + 
  geom_point(alpha = 0.2)  +
  # geom_line() + 
  facet_grid(rca~year.f, scales = "free")
```


```{r eval = F}
hourly2 %>% filter (month == 4) %>%
ggplot(., aes(datehr, gamresid)) + 
  geom_point(alpha = 0.2)  +
  geom_line() +
  facet_grid(rca~year.f, scales = "free")
```
